Had the search view forward to the list as well, which now has a few more items. Didn't get to any location stuff.
Current bug: Search passes
intent to SearchableActivity which then immediately creates an intent
which it uses to forward to the ListView(ResultsActivity). Using the
back button from the ListView goes to the blank SearchableActivity
layout though. Should probably go straight back to the main screen but
we'll roll with it for demonstrations sake.